Changed the interface of ParentChildCreatorInterface: The create method no longer takes a Parameter argument but a more general Map. This way the action can pass in arbitrary Objects to the create method and ultimatively to the transformXML methods.
The action now grabs all request, sitemap and session parameter and
fills them into a map which is then passed along to the create method
which hands it down to the transform method. This way the transform
method can fill, e.g. the user (from the session) into a new document.


git-svn-id: https://svn.apache.org/repos/asf/incubator/lenya/trunk@38912 13f79535-47bb-0310-9956-ffa450edef68