Image upload for uni public works again. The following changes had to be applied:
The sitemap path was not pointing at the publication sitemap
anymore since the usecase stuff has been moved to the root
sitemap. So the PageEnvelope is used to find the publication
directory.

In the olden days (when uni public was created) the
document-id used to contain the suffix of the document as
well. This is no longer the case, so the image upload had to
be adjusted.


git-svn-id: https://svn.apache.org/repos/asf/incubator/lenya/trunk@40183 13f79535-47bb-0310-9956-ffa450edef68