added ablity to call module resources while specifying the publication id
Now module resources and sitemap can be referenced via:
http://localhost:8888/modules/{module-id}/...
or
http://localhost:8888/{publication-id}/modules/{module-id}/...

Both will return the same thing if accessed as an internal request.
For external requests the first will always return core resources, while the second 
will be able to utilize the fallback capability of module-resources.xmap or {module-id}/sitemap.xmap.

git-svn-id: https://svn.apache.org/repos/asf/lenya/trunk@376020 13f79535-47bb-0310-9956-ffa450edef68