ASF Bugzilla Bug 39906 - Move Xopus integration code to a module From Jonathan Addison. Thank you.
This patch moves all of the xopus files and pipelines from the core to a new
xopus module.

It also fixes some issues left over from migrating xopus from 1.2 to 1.4:
- deleted old stylesheets
- added xopus.path parameter next to xopus.context parameter in xopus.xconf
(previously the pipeline had to be updated with the relative path of the xopus
location)
- added a basic xhtml schema (thanks to Josias)
- added strip_namespaces.xsl to the xopus transformation pipeline

There is a README.txt that covers what needs to be set up.

git-svn-id: https://svn.apache.org/repos/asf/lenya/trunk@422748 13f79535-47bb-0310-9956-ffa450edef68