This commit uses a generator of a forrest plugin. See http://svn.apache.org/viewvc?view=rev&revision=429872 for more information.
Produces properties from an input-module. This is been managed by the 
heart of the forrest plugin the ModuleGenerator. The generator is contacting 
the input module and gets an iterator of all attributes names. This 
iterator will be then used to receive the values of the key. Till now 
we only use a Object to String cast but an input module does not have 
to return a simple String but can return any object. This may cause 
ClassCastExceptions for the values.

Further some attributes need to be received with additional parameter 
(like the page-envelope). The genarator returns then Null.

See 
http://localhost:8888/modules/properties/input-module.properties.properties.xml 
http://localhost:8888/modules/properties/input-module.page-envelope.properties.xml

The match is
 <map:match pattern="input-module.*.properties.xml">

git-svn-id: https://svn.apache.org/repos/asf/lenya/trunk@431181 13f79535-47bb-0310-9956-ffa450edef68