create continuations for usecase uri views (to make them consistent with jxtemplate views)
uri views (i.e. views defined with <view uri="..." .../> in the
usecase.xconf) used to be one-shot cocoon.sendPage beasts. no
continuation was created for them, as opposed to <view template="..."
.../> usecases, for which the usecase framework would call
cocoon.sendPageAndWait. 
this meant that the mechanism of the view fundamentally affected the
behaviour of the controller, which is simply wrong in an MVC context.

currently, only three usecases in trunk are affected by this change:
* bxe.open
* bxe.close
* jcrsource.export
the bxe stuff works as ever, but iiuc a stale continuation will be
created and the usecase will never execute and finish. that is no
problem, since the continuation will eventually time out, but it should
be fixed at some point.
i could not test the behaviour of the jcrsource module.



git-svn-id: https://svn.apache.org/repos/asf/lenya/trunk@448941 13f79535-47bb-0310-9956-ffa450edef68