make bxe work in proxied environments. the primary problem was that bxe needs an rng file on the client, and we did not correctly resolve <rng:include href=".."> urls.
there are two ugly hacks in the existing code that should
eventually be removed altogether. the current patch just adds
workarounds, since we're in code freeze.
the hacks are:
* webapp/lenya/xslt/resources/translate-rng-includes.xsl implements a
fallback uri resolver for client requests and gets it wrong. the error
has been fixed and a proxy transformation added, but we should not duplicate knowledge about fallback
resolving in an xslt! 
* java/org/apache/lenya/cms/cocoon/components/modules/input/ResourceTypeModule.java
adds a context prefix to the schema uri and also tries to resolve
fallback itself, again not really watertight. the context prefix has
been removed, and we now apply a proxy transformer afterwards.

to make this work, the proxy transformer configuration was extended to catch
rng:include elements.

TODO: bxe.close still redirects to a wrong URL (context prefix instead
of proxy)


git-svn-id: https://svn.apache.org/repos/asf/lenya/trunk@594615 13f79535-47bb-0310-9956-ffa450edef68